[#691] add filters to results page for mobile
