+++ src/components/Menubar/useMobileSearch.ts
@@ -1,74 +1,28 @@
 import { useAppDispatch, useAppSelector } from '@/app/hooks'
 import { selectCalendars } from '@/app/selectors/selectCalendars'
 import { Calendar } from '@/features/Calendars/CalendarTypes'
-import { searchEventsAsync } from '@/features/Search/SearchSlice'
+import {
+    clearFilters,
+    searchEventsAsync, SearchFilters, setFilters
+} from '@/features/Search/SearchSlice'
+import { buildQuery } from '@/features/Search/searchUtils'
 import { setView } from '@/features/Settings/SettingsSlice'
-import { userAttendee } from '@/features/User/models/attendee'
 import { createAttendee } from '@/features/User/models/attendee.mapper'
 import { extractEventBaseUuid } from '@/utils/extractEventBaseUuid'
 import { useCallback, useMemo, useState } from 'react'
 import { User } from '../Attendees/types'
 import { SearchState } from '../Calendar/utils/tempSearchUtil'
 
-type Filters = {
-  searchIn: string
-  keywords: string
-  organizers: userAttendee[]
-  attendees: userAttendee[]
-}
-
-function getSearchInCalendars(
-  searchIn: string,
-  allIds: string[],
-  personalIds: string[]
-): string[] {
-  if (!searchIn) return allIds
-  if (searchIn === 'my-calendars') return personalIds
-  return [searchIn]
-}
-
-function buildQuery(
-  searchQuery: string,
-  filters: Filters,
-  allIds: string[],
-  personalIds: string[]
-):
-  | {
-      search: string
-      filters: {
-        keywords: string
-        organizers: string[]
-        attendees: string[]
-        searchIn: string[]
-      }
-    }
-  | undefined {
-  const trimmedSearch = searchQuery.trim()
-  const trimmedKeywords = filters.keywords.trim()
-  const hasSearchCriteria =
-    trimmedSearch ||
-    trimmedKeywords ||
-    filters.organizers.length > 0 ||
-    filters.attendees.length > 0
-  if (!hasSearchCriteria) return undefined
-  return {
-    search: trimmedSearch,
-    filters: {
-      keywords: trimmedKeywords,
-      organizers: filters.organizers.map(u => u.cal_address),
-      attendees: filters.attendees.map(u => u.cal_address),
-      searchIn: getSearchInCalendars(filters.searchIn, allIds, personalIds)
-    }
-  }
-}
-
 export function useMobileSearch(setDialogOpen: (b: boolean) => void): {
   inputQuery: string
   setInputQuery: React.Dispatch<React.SetStateAction<string>>
   searchState: SearchState
   selectedContacts: User[]
-  filters: Filters
-  handleSearch: (searchQuery: string, currentFilters: Filters) => Promise<void>
+  filters: SearchFilters
+  handleSearch: (
+    searchQuery: string,
+    currentFilters: SearchFilters
+  ) => Promise<void>
   handleSearchChange: ({ query, options, loading }: SearchState) => void
   handleContactSelect: (contacts: User[]) => void
   clearAll: () => void
@@ -77,6 +31,10 @@ export function useMobileSearch(setDialogOpen: (b: boolean) => void): {
   const dispatch = useAppDispatch()
   const calendars = useAppSelector(selectCalendars)
   const userId = useAppSelector(state => state.user.userData?.openpaasId)
+  const filters = useAppSelector(
+    state => state.searchResult.searchParams.filters
+  )
+
   const personalCalendars = useMemo(
     (): Calendar[] =>
       userId
@@ -92,15 +50,11 @@ export function useMobileSearch(setDialogOpen: (b: boolean) => void): {
     loading: false
   })
   const [selectedContacts, setSelectedContacts] = useState<User[]>([])
-  const [filters, setFilters] = useState<Filters>({
-    searchIn: 'my-calendars',
-    keywords: '',
-    organizers: [],
-    attendees: []
-  })
-
   const handleSearch = useCallback(
-    async (searchQuery: string, currentFilters: Filters): Promise<void> => {
+    async (
+      searchQuery: string,
+      currentFilters: SearchFilters
+    ): Promise<void> => {
       const query = buildQuery(
         searchQuery,
         currentFilters,
@@ -139,7 +93,7 @@ export function useMobileSearch(setDialogOpen: (b: boolean) => void): {
       )
       const nextFilters = { ...filters, organizers }
       setSelectedContacts(contacts)
-      setFilters(nextFilters)
+      dispatch(setFilters(nextFilters))
       setInputQuery('')
       setSearchState({ query: '', options: [], loading: false })
       if (contacts.length > 0) {
@@ -153,12 +107,7 @@ export function useMobileSearch(setDialogOpen: (b: boolean) => void): {
     setInputQuery('')
     setSelectedContacts([])
     setSearchState({ query: '', options: [], loading: false })
-    setFilters(prev => ({
-      ...prev,
-      keywords: '',
-      attendees: [],
-      organizers: []
-    }))
+    dispatch(clearFilters())
     setDialogOpen(false)
   }, [setDialogOpen])
