+++ src/features/Search/searchUtils.ts
@@ -1,0 +1,58 @@
+import { userAttendee } from '@/features/User/models/attendee'
+import { SearchFilters } from './SearchSlice'
+
+export function buildOrganizersAndAttendeesFilter(filters: {
+  searchIn: string
+  keywords: string
+  organizers: userAttendee[]
+  attendees: userAttendee[]
+}): string[] {
+  return filters.organizers.map(u => u.cal_address)
+}
+
+export function getSearchInCalendars(
+  searchIn: string,
+  allIds: string[],
+  personalIds: string[]
+): string[] {
+  if (!searchIn) return allIds
+  if (searchIn === 'my-calendars') return personalIds
+  return [searchIn]
+}
+
+export function buildQuery(
+  searchQuery: string,
+  filters: SearchFilters,
+  allIds: string[],
+  personalIds: string[]
+):
+  | {
+      search: string
+      filters: {
+        searchIn: string[]
+        keywords: string
+        organizers: string[]
+        attendees: string[]
+      }
+    }
+  | undefined {
+  const trimmedSearch = searchQuery.trim()
+  const trimmedKeywords = filters.keywords.trim()
+  const hasSearchCriteria =
+    trimmedSearch ||
+    trimmedKeywords ||
+    filters.organizers.length > 0 ||
+    filters.attendees.length > 0
+
+  if (!hasSearchCriteria) return undefined
+
+  return {
+    search: trimmedSearch,
+    filters: {
+      keywords: trimmedKeywords,
+      organizers: filters.organizers.map(u => u.cal_address),
+      attendees: filters.attendees.map(u => u.cal_address),
+      searchIn: getSearchInCalendars(filters.searchIn, allIds, personalIds)
+    }
+  }
+}
